# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThankATech API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Technician {
  id: number;
  fullName: string;
  /** @nullable */
  avatarUrl: string | null;
  specialty: string;
  specialties?: string[];
  serviceArea: string;
  bio: string;
  /** @nullable */
  hourlyRate?: number | null;
  certifications?: string[];
  totalThanks: number;
  totalEarned: number;
  createdAt?: string;
}

export interface TechnicianInput {
  fullName: string;
  avatarUrl?: string;
  specialty: string;
  specialties?: string[];
  serviceArea: string;
  bio: string;
  hourlyRate?: number;
  certifications?: string[];
}

export interface TechnicianStats {
  technicianId: number;
  totalThanks: number;
  totalTips: number;
  totalEarned: number;
  totalJobs: number;
  avgTipAmount: number;
}

export interface Job {
  id: number;
  customerId: number;
  customerName?: string;
  technicianId: number;
  technicianName?: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  address?: string | null;
  status: string;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface JobInput {
  customerId: number;
  technicianId: number;
  title: string;
  description?: string;
  address?: string;
}

export interface JobUpdate {
  status?: string;
  completedAt?: string;
}

export interface ThankMessage {
  id: number;
  jobId: number;
  customerId: number;
  customerName?: string;
  technicianId: number;
  technicianName?: string;
  /** @nullable */
  technicianAvatar?: string | null;
  message: string;
  tipAmount: number;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface ThankMessageInput {
  jobId: number;
  customerId: number;
  technicianId: number;
  message: string;
  tipAmount?: number;
  photoUrl?: string;
}

export interface PointsBalance {
  userId: number;
  balance: number;
  updatedAt?: string;
}

export interface PointTransaction {
  id: number;
  userId: number;
  amount: number;
  type: string;
  /** @nullable */
  jobId?: number | null;
  description?: string;
  createdAt: string;
}

export interface PlatformStats {
  totalTechnicians: number;
  totalThanks: number;
  totalTipsAmount: number;
  totalJobs: number;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface MyProfile {
  profileId: number;
  userType: string;
  /** @nullable */
  technicianId?: number | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface MyProfileEnvelope {
  profile: MyProfile | null;
}

export interface UpsertProfileInput {
  userType: string;
  fullName: string;
  specialty?: string;
  serviceArea?: string;
  bio?: string;
}

export interface StripeConfig {
  publishableKey: string;
}

export interface StripeOnboardingUrl {
  url: string;
}

export interface StripeConnectStatus {
  connected: boolean;
  onboardingComplete: boolean;
  /** @nullable */
  accountId?: string | null;
}

export interface StripePaymentIntentInput {
  thankMessageId: number;
  amount: number;
}

export interface StripePaymentIntentResult {
  clientSecret: string;
  paymentIntentId: string;
  connectedToStripe: boolean;
}

export interface StripePaymentCompleteInput {
  thankMessageId: number;
  paymentIntentId: string;
}

export interface StripePaymentCompleteResult {
  success: boolean;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListTechniciansParams = {
  specialty?: string;
  search?: string;
};

export type ListJobsParams = {
  customerId?: number;
  technicianId?: number;
  status?: string;
};

export type ListThankMessagesParams = {
  technicianId?: number;
  customerId?: number;
};

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};