# lib/api-zod/src/generated/types/stripeConnectStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThankATech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StripeConnectStatus {
  connected: boolean;
  onboardingComplete: boolean;
  /** @nullable */
  accountId?: string | null;
}