# lib/api-zod/src/generated/types/stripePaymentCompleteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThankATech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StripePaymentCompleteResult {
  success: boolean;
}