# lib/api-zod/src/generated/types/stripePaymentIntentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThankATech API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StripePaymentIntentResult {
  clientSecret: string;
  paymentIntentId: string;
  connectedToStripe: boolean;
}